Reorganize into multiple files
